Rename index.html to home.html
Changed the name of the homepage HTML file from 'index.html' to 'home.html' to stop automatic serving of the homepage. This is in an effort to take down the site (rather than refactoring the theme/fonts) by direction of University Marketing as the Verlag font used in the UW theme license expired as of July 1, 2023.